A working B failing on large input
LZW-A properly encodes for all tested cases in corpus calgary.
Obviously compression is negative in almost all cases, but it was
written as a stepping stone to further implementations of the LZW
compression method.

LZW-b works, except for two things: it truncates the last character of
the input string, and it fails on extremely large inputs.  The failing
is due to the fact that 2 byte integers are used as codes--we start to
get junk data after ~65000 codes have been used.